[TECH-293] Extract collection class to separate module
branch: feature/TECH-293-customizable-stages
